Fix cross-platform launcher quoting, terminal/libterminal loading, and Lua 5.4 unpack compat

+++ src/api/renderer.c
@@ -249,7 +249,7 @@ static int f_font_get_size(lua_State *L) {
 static int f_font_set_size(lua_State *L) {
   RenFont* fonts[FONT_FALLBACK_MAX]; font_retrieve(L, fonts, 1);
   float size = luaL_checknumber(L, 2);
-  int scale = 1;
+  double scale = 1;
 #ifdef LITE_USE_SDL_RENDERER
   RenWindow *window = ren_get_target_window();
   if (window != NULL) {
